+++ src/main/java/com/dropcube/beans/mocks/Datum__.java
@@ -7,7 +7,7 @@ public class Datum__ {
 
     @SerializedName("time")
     @Expose
-    private Integer time;
+    private Long time;
     @SerializedName("summary")
     @Expose
     private String summary;
@@ -87,11 +87,11 @@ public class Datum__ {
     @Expose
     private Double ozone;
 
-    public Integer getTime() {
+    public Long getTime() {
         return time;
     }
 
-    public void setTime(Integer time) {
+    public void setTime(Long time) {
         this.time = time;
     }
